moved hooks and utils to the proper dir
