+++ packages/suite-base/src/panels/diagnostics/hooks/useAvailableDiagnostics.ts
