+++ packages/suite-base/src/panels/diagnostics/hooks/useStaleTime.ts
