+++ packages/suite-base/src/panels/diagnostics/utils/settings.ts
