+++ packages/suite-base/src/panels/diagnostics/utils/util.test.ts
