+++ packages/suite-base/src/panels/diagnostics/utils/util.ts
